backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
greenlet>=3.0.0
python-dotenv>=1.0.0
openai>=1.55.0
langgraph>=0.2.60
langchain-core>=0.3.0
httpx>=0.27.0
PyJWT>=2.8.0
edge-tts>=7.0.0
mutagen>=1.47.0
